+++ R/kernels.R
@@ -1,12 +1,11 @@
-krbf <- function(X, Y = NULL, sigma = 1) {
-    exp(-.5 * euclsq(X, Y) / sigma^2)
-    }
+krbf <- function(X, Y = NULL, gamma = 1)
+    exp(-gamma * euclsq(X, Y))
 
-kpol <- function(X, Y = NULL, degree = 1, scale = 1, offset = 0) {
+kpol <- function(X, Y = NULL, degree = 1, gamma = 1, coef0 = 0) {
     if(is.null(Y))
-        K <- (scale * tcrossprod(X) + offset)
+        K <- (gamma * tcrossprod(X) + coef0)
     else
-        K <- (scale * tcrossprod(X, Y) + offset)
+        K <- (gamma * tcrossprod(X, Y) + coef0)
     if(degree > 1) {
         zK <- K
         for(i in seq_len(degree - 1))
@@ -15,8 +14,8 @@ kpol <- function(X, Y = NULL, degree = 1, scale = 1, offset = 0) {
     K    
     }
 
-ktanh <- function(X, Y = NULL, scale = 1, offset = 0)
-    tanh(kpol(X, Y, degree = 1, scale, offset))
+ktanh <- function(X, Y = NULL, gamma = 1, coef0 = 0)
+    tanh(kpol(X, Y, degree = 1, gamma, coef0))
 
 #kgram <- function(Xtrain, X = NULL, kern = krbf, ...) {
 #    Xtrain <- .mat(Xtrain)